+++ demos/parse-server-migration/src/index.ts
@@ -6,6 +6,7 @@ import { parseServer } from './parseServer';
 // @ts-ignore
 import ParseServer from 'parse-server';
 import http from 'http';
+import { streamsSync } from '@moralisweb3/parse-server';
 
 export const app = express();
 
@@ -18,7 +19,14 @@ app.use(express.json());
 
 app.use(cors());
 
-app.use(`/server`, parseServer);
+app.use(
+  streamsSync(parseServer, {
+    apiKey: config.MORALIS_API_KEY,
+    webhookUrl: '/streams',
+  }),
+);
+
+app.use(`/server`, parseServer.app);
 
 const httpServer = http.createServer(app);
 httpServer.listen(config.PORT, () => {