+++ packages/streams/src/resolvers/evmStreams/addAddressEvm.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
-import { toCamelCase } from '@moralisweb3/common-core';
-import { EvmAddress } from '@moralisweb3/common-evm-utils';
+import { toCamelCase } from '@moralisweb3/core';
+import { EvmAddress, EvmAddressish } from '@moralisweb3/evm-utils';
 import { operations } from '../../generated/types';
 
 const name = 'AddAddressToStream';
@@ -10,18 +10,18 @@ type PathParams = operations[Name]['parameters']['path'];
 type BodyParams = operations[Name]['requestBody']['content']['application/json'];
 // Overwriting "address: string | string[]" because the generator cannot infer the type correctly from swagger
 type ApiParams = Omit<PathParams & BodyParams, 'address'> & { address: string | string[] };
-export type AddAddressEvmParams = ApiParams;
+export interface AddAddressEvmParams extends Omit<ApiParams, 'address'> {
+  address: EvmAddressish | EvmAddressish[];
+}
 const method = 'post';
 const bodyParams = ['address'] as const;
-const urlParams = ['id'] as const;
 
 type ApiResult = operations[Name]['responses']['200']['content']['application/json'];
 
 export const addAddressEvm = createEndpointFactory(() =>
   createEndpoint({
     name,
-    getUrl: ({ id }: AddAddressEvmParams) => `/streams/evm/${id}/address`,
-    urlParams,
+    getUrl: (params: AddAddressEvmParams) => `/streams/evm/${params.id}/address`,
     apiToResult: (apiData: ApiResult) => {
       const data = toCamelCase(apiData);
 
@@ -42,7 +42,12 @@ export const addAddressEvm = createEndpointFactory(() =>
           : data.address.format()
         : undefined,
     }),
-    parseParams: (params: AddAddressEvmParams): ApiParams => params,
+    parseParams: (params: AddAddressEvmParams): ApiParams => ({
+      ...params,
+      address: Array.isArray(params.address)
+        ? params.address.map((address) => EvmAddress.create(address).lowercase)
+        : EvmAddress.create(params.address).lowercase,
+    }),
     method,
     bodyParams,
   }),